+++ omc3/segment_by_segment/definitions.py
@@ -19,7 +19,7 @@ class PropagableColumns:
     """ Class to define columns for propagables. 
     One could also implicitly define the error-columns,
     either via __getattr__ or as a wrapper, but I decided to
-    explicityly name these columns, so that the IDEs can see them 
+    explicitely name these columns, so that the IDEs can see them 
     and can help in renaming and autocompletion (jdilly 2023).
     """
     def __init__(self, column: str, plane: str = "{}") -> None: