+++ tests/accuracy/test_response_creator.py
@@ -4,7 +4,6 @@ import pytest
 from omc3.correction.response_io import (
     read_fullresponse,
     read_varmap,
-    write_fullresponse,
 )
 from omc3.correction.sequence_evaluation import evaluate_for_variables
 from omc3.global_correction import OPTICS_PARAMS_CHOICES
@@ -14,7 +13,6 @@ from omc3.response_creator import create_response_entrypoint as create_response
 from omc3.utils import logging_tools
 from omc3.utils.stats import rms
 from tests.accuracy.test_global_correction import get_normal_params, get_skew_params
-from tests.conftest import INPUTS
 
 LOG = logging_tools.get_logger(__name__)
 # LOG = logging_tools.get_logger('__main__', level_console=logging_tools.MADX)